Decouple lattice properties from interface for STQBM
