+++ qlbm/components/base.py
@@ -6,7 +6,8 @@ from qiskit import QuantumCircuit
 from qiskit.qasm2 import dump as dump_qasm2
 from qiskit.qasm3 import dump as dump_qasm3
 
-from qlbm.lattice import CollisionlessLattice, Lattice, SpaceTimeLattice
+from qlbm.lattice import CollisionlessLattice, Lattice
+from qlbm.lattice.lattices.spacetime.spacetime_lattice import SpaceTimeLattice
 
 
 class QuantumComponent(ABC):
